Clean-up: Rename source file.
